В GraphViewModel.kt доработаны функции для загрузки зав. номеров и букв. кодов датчиков, имеющихся в БД.

+++ data/src/main/java/ru/profitsw2000/data/model/state/sensorfilterscreen/SensorIdsLoadState.kt
@@ -2,6 +2,6 @@ package ru.profitsw2000.data.model.state.sensorfilterscreen
 
 sealed class SensorIdsLoadState {
     data object Loading: SensorIdsLoadState()
-    data class Success(val sensorIdsList: List<Pair<Int, Boolean>>): SensorIdsLoadState()
+    data class Success(val sensorIdsList: List<Pair<Long, Boolean>>): SensorIdsLoadState()
     data object Error: SensorIdsLoadState()
 }
